+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The AdLer Engine uses [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- fixed only showing short break time content
+
 ### Deprecated
 
 ### Security